refactor: Add Appium as peer dependency
