+++ .azure-pipelines.yml
@@ -1,10 +1,10 @@
 # https://github.com/Microsoft/azure-pipelines-image-generation/blob/master/images/macos/
 variables:
-  VM_IMAGE: macOS-10.15
-  XCODE_VERSION: 12.1
-  PLATFORM_VERSION: 14.1
-  DEVICE_NAME: iPhone 11 Pro Max
-  DEFAULT_NODE_VERSION: "12.x"
+  VM_IMAGE: macOS-12
+  XCODE_VERSION: 13.4
+  PLATFORM_VERSION: 15.5
+  DEVICE_NAME: iPhone 12
+  DEFAULT_NODE_VERSION: "16.x"
 
 pool:
   vmImage: "$(VM_IMAGE)"
