+++ lib/commands/record-screen.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { util, fs, net, tempDir } from '@appium/support';
+import { util, fs, net, tempDir } from 'appium/support';
 import { waitForCondition } from 'asyncbox';
 import Simctl from 'node-simctl';
