+++ lib/driver.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { BaseDriver } from '@appium/base-driver';
+import { BaseDriver } from 'appium/driver';
 import SafariDriverServer from './safari';
 import { desiredCapConstraints } from './desired-caps';
 import commands from './commands/index';
