+++ lib/safari.js
@@ -1,7 +1,7 @@
 import os from 'os';
 import path from 'path';
-import { JWProxy, errors } from '@appium/base-driver';
-import { fs, logger, util } from '@appium/support';
+import { JWProxy, errors } from 'appium/driver';
+import { fs, logger, util } from 'appium/support';
 import { SubProcess } from 'teen_process';
 import { waitForCondition } from 'asyncbox';
 import { findAPortNotInUse } from 'portscanner';
