+++ package.json
@@ -33,8 +33,6 @@
     "lib": "lib"
   },
   "dependencies": {
-    "@appium/base-driver": "^8.3.0",
-    "@appium/support": "^2.46.0",
     "@babel/runtime": "^7.0.0",
     "asyncbox": "^2.0.2",
     "bluebird": "^3.5.1",
@@ -62,12 +60,17 @@
     "precommit-msg",
     "precommit-test"
   ],
+  "peerDependencies": {
+    "appium": "^2.0.0-beta.40"
+  },
   "devDependencies": {
     "@appium/gulp-plugins": "^7.0.0",
     "@appium/eslint-config-appium": "^6.0.0",
     "@babel/core": "^7.16.0",
     "@babel/eslint-parser": "^7.16.3",
     "@semantic-release/git": "^10.0.1",
+    "appium": "^2.0.0-beta.40",
+    "appium-safari-driver": "file:.",
     "chai": "^4.1.2",
     "chai-as-promised": "^7.1.1",
     "eslint-config-prettier": "^8.5.0",
