+++ test/functional/desktop-driver-e2e-specs.js
@@ -1,5 +1,5 @@
 import { remote } from 'webdriverio';
-import { startServer } from '../..';
+import { startServer } from '../server';
 import chaiAsPromised from 'chai-as-promised';
 import chai from 'chai';
 import { HOST, PORT, MOCHA_TIMEOUT } from '../utils';
@@ -11,6 +11,7 @@ chai.use(chaiAsPromised);
 const CAPS = {
   browserName: 'AppiumSafari',
   platformName: 'mac',
+  'appium:automationName': 'safari',
 };
 
 describe('Desktop SafariDriver', function () {
